requirements.txt
click
matplotlib
numpy
python-dotenv
scikit_learn
setuptools
torch
torchvision
